+++ tests/attacks/test_zoo.py
@@ -165,7 +165,8 @@ class TestZooAttack(unittest.TestCase):
         # Build PyTorchClassifier
         ptc = get_classifier_pt()
 
-        x_test = np.swapaxes(self.x_test, 1, 3)
+        # Get MNIST
+        x_test = np.swapaxes(self.x_test, 1, 3).astype(np.float32)
 
         # First attack
         # zoo = ZooAttack(classifier=ptc, targeted=True, max_iter=10, binary_search_steps=10)
